+++ packages/vsce/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 ## Recent Changes
 
 - Validate input using maximum resource length when inspecting resource. [#402](https://github.com/zowe/cics-for-zowe-client/issues/402)
+- Action menus on Resource Inspector Header. [#333](https://github.com/zowe/cics-for-zowe-client/issues/333)
 
 ## `3.11.0`